requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiofiles==0.4.0           # via -r requirements.in (line 2), fastapi
aiohttp==3.6.2            # via -r requirements.in (line 3)
alembic==1.4.0            # via -r requirements.in (line 4)
aniso8601==7.0.0          # via graphene
arrow==0.15.5             # via -r requirements.in (line 5), rq-dashboard
async-exit-stack==1.0.1   # via fastapi
async-generator==1.10     # via fastapi
async-timeout==3.0.1      # via aiohttp
asyncpg==0.20.1           # via databases
attrs==19.3.0             # via aiohttp
bcrypt==3.1.7             # via -r requirements.in (line 6)
certifi==2019.11.28       # via requests
cffi==1.14.0              # via bcrypt
chardet==3.0.4            # via aiohttp, requests
click-log==0.3.2          # via -r requirements.in (line 7)
click==7.0                # via -r requirements.in (line 8), click-log, flask, rq, uvicorn
croniter==0.3.31          # via rq-scheduler
databases[postgresql]==0.2.6  # via -r requirements.in (line 11)
decorator==4.4.1          # via traitlets
dnspython==1.16.0         # via email-validator
email-validator==1.0.5    # via fastapi, pydantic
fastapi[all]==0.49.1      # via -r requirements.in (line 10)
flask==1.1.1              # via rq-dashboard
geoalchemy2==0.6.3        # via -r requirements.in (line 9)
graphene==2.1.8           # via fastapi
graphql-core==2.3.1       # via graphene, graphql-relay
graphql-relay==2.0.1      # via graphene
h11==0.9.0                # via uvicorn
httptools==0.1.1          # via uvicorn
idna==2.9                 # via email-validator, requests, yarl
invoke==1.4.1             # via -r requirements.in (line 12)
ipython-genutils==0.2.0   # via traitlets
itsdangerous==1.1.0       # via fastapi, flask
jinja2==2.11.1            # via -r requirements.in (line 13), fastapi, flask
loguru==0.4.1             # via -r requirements.in (line 14)
mako==1.1.1               # via alembic
markupsafe==1.1.1         # via jinja2, mako
multidict==4.7.5          # via aiohttp, yarl
numpy==1.18.1             # via -r requirements.in (line 15), pandas
pandas==1.0.1             # via -r requirements.in (line 16)
promise==2.3              # via graphql-core, graphql-relay
psycopg2-binary==2.8.4    # via databases
pycparser==2.19           # via cffi
pydantic[dotenv,email,typing_extensions]==1.4  # via -r requirements.in (line 31), fastapi
pyjwt==1.7.1              # via -r requirements.in (line 17)
python-dateutil==2.8.1    # via -r requirements.in (line 18), alembic, arrow, croniter, pandas
python-dotenv==0.12.0     # via -r requirements.in (line 19), pydantic
python-editor==1.0.4      # via alembic
python-multipart==0.0.5   # via fastapi
pytz==2019.3              # via -r requirements.in (line 20), pandas
pyyaml==5.3               # via -r requirements.in (line 21), fastapi
redis==3.4.1              # via -r requirements.in (line 22), rq, rq-dashboard
requests==2.23.0          # via fastapi
rq-dashboard==0.6.1       # via -r requirements.in (line 23)
rq-scheduler==0.9.1       # via -r requirements.in (line 24)
rq==1.2.2                 # via -r requirements.in (line 25), rq-dashboard, rq-scheduler
rx==1.6.1                 # via graphql-core
six==1.14.0               # via bcrypt, graphene, graphql-core, graphql-relay, python-dateutil, python-multipart, sqlalchemy-utils, traitlets
sqlalchemy-utils==0.36.1  # via -r requirements.in (line 26)
sqlalchemy==1.3.13        # via -r requirements.in (line 27), alembic, databases, geoalchemy2, sqlalchemy-utils
starlette==0.12.9         # via fastapi
traitlets==4.3.3          # via -r requirements.in (line 28)
typing-extensions==3.7.4.1  # via -r requirements.in (line 29), pydantic
ujson==1.35               # via fastapi
urllib3==1.25.8           # via requests
uvicorn==0.11.3           # via -r requirements.in (line 30), fastapi
uvloop==0.14.0            # via uvicorn
websockets==8.1           # via uvicorn
werkzeug==1.0.0           # via flask
yarl==1.4.2               # via aiohttp
